[refactor] filter -> selectMenu 파일명 및 폴더명 변경 - 더 직관적인 네이밍을 위해 파일명 변경 - selectMenu -> selectMenuModal -> selectMenuItem
