+++ client/src/components/atoms/selectMenuItem/ColorSwatch/ColorSwatch.tsx
