+++ client/src/components/atoms/selectMenuItem/Image/index.ts
