--- client/src/components/molecules/FilterModal/FilterModalButton/filterItem/FilterItemLabel/index.ts
@@ -1,1 +1,0 @@
-export { default } from './FilterItemLabel';
