--- client/src/components/molecules/FilterModal/FilterModalButton/index.ts
@@ -1,1 +1,0 @@
-export { default } from './FilterModalButton';
