--- client/src/components/molecules/FilterModal/index.ts
@@ -1,1 +1,0 @@
-export { default } from './FilterModal';
