+++ client/src/components/molecules/SelectMenuModal/SelectMenuItem/SelectMenuItemLabel/index.ts
@@ -1,0 +1,1 @@
+export { default } from './SelectMenuItemLabel';
