+++ client/src/components/molecules/SelectMenuModal/SelectMenuItem/SelectMenuItemUser/SelectMenuItemUser.stories.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
 import { text } from '@storybook/addon-knobs';
-import FilterItemUser from './FilterItemUser';
+import SelectMenuItemUser from './SelectMenuItemUser';
 
 export default {
-  title: 'Molecules/FilterItemUser',
-  component: FilterItemUser,
+  title: 'Molecules/SelectMenuItemUser',
+  component: SelectMenuItemUser,
 };
 
 export const Default = () => {
@@ -13,5 +13,5 @@ export const Default = () => {
     'https://avatars0.githubusercontent.com/u/37282087?s=40&amp;v=4';
   const alt = 'user';
 
-  return <FilterItemUser title={title} userImgSrc={userImgSrc} alt={alt} />;
+  return <SelectMenuItemUser title={title} userImgSrc={userImgSrc} alt={alt} />;
 };
