+++ client/src/components/molecules/SelectMenuModal/SelectMenuItem/SelectMenuItemUser/index.ts
@@ -1,0 +1,1 @@
+export { default } from './SelectMenuItemUser';
