+++ client/src/components/molecules/SelectMenuModal/SelectMenuItem/index.ts
@@ -1,0 +1,1 @@
+export { default } from './SelectMenuItem';
