+++ client/src/components/molecules/SelectMenuModal/index.ts
@@ -1,0 +1,1 @@
+export { default } from './SelectMenuModal';
