--- client/src/components/molecules/filterItem/FilterItemUser/index.ts
@@ -1,1 +1,0 @@
-export { default } from './FilterItemUser';
